refactor: add footer component
